Change Point struct to use u32 instead of i16

+++ src/base.zig
@@ -122,8 +122,8 @@ pub const Mouse = enum {
 };
 
 pub const Point = struct {
-    x: i16,
-    y: i16,
+    x: u32,
+    y: u32,
 };
 
 pub const Rect = struct {
